--- packages/cra-template-safe-app/src/App.test.tsx
@@ -1,8 +1,0 @@
-import { render, screen } from './tests/utils';
-import App from './App';
-
-test('renders learn loading screen', () => {
-  render(<App />);
-  const waitingHeading = screen.getByText(/Waiting for Safe/i);
-  expect(waitingHeading).toBeInTheDocument();
-});
