--- packages/cra-template-safe-app/src/index.tsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { ThemeProvider } from 'styled-components';
-import { theme, Loader, Title } from '@gnosis.pm/safe-react-components';
-import SafeProvider from '@gnosis.pm/safe-apps-react-sdk';
-
-import GlobalStyle from './GlobalStyle';
-import App from './App';
-
-ReactDOM.render(
-  <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <GlobalStyle />
-      <SafeProvider
-        loader={
-          <>
-            <Title size="md">Waiting for Safe...</Title>
-            <Loader size="md" />
-          </>
-        }
-      >
-        <App />
-      </SafeProvider>
-    </ThemeProvider>
-  </React.StrictMode>,
-  document.getElementById('root'),
-);
