+++ packages/cra-template-safe-app/template/src/App.test.tsx
@@ -1,8 +1,8 @@
-import { render, screen } from "@testing-library/react"
-import App from "./App"
+import { render, screen } from './tests/utils';
+import App from './App';
 
-test("renders learn loading screen", () => {
-  render(<App />)
-  const waitingHeading = screen.getByText(/Waiting for Safe/i)
-  expect(waitingHeading).toBeInTheDocument()
-})
+test('renders learn loading screen', () => {
+  render(<App />);
+  const waitingHeading = screen.getByText(/Waiting for Safe/i);
+  expect(waitingHeading).toBeInTheDocument();
+});
