+++ packages/cra-template-safe-app/template/src/App.tsx
@@ -1,14 +1,14 @@
-import React, { useCallback } from "react"
-import styled from "styled-components"
-import { Button, Title } from "@gnosis.pm/safe-react-components"
-import { useSafeAppsSDK } from "@gnosis.pm/safe-apps-react-sdk"
+import React, { useCallback } from 'react';
+import styled from 'styled-components';
+import { Button, Title } from '@gnosis.pm/safe-react-components';
+import { useSafeAppsSDK } from '@gnosis.pm/safe-apps-react-sdk';
 
 const Container = styled.div`
   padding: 1rem;
-`
+`;
 
 const SafeApp = (): React.ReactElement => {
-  const { sdk, safe } = useSafeAppsSDK()
+  const { sdk, safe } = useSafeAppsSDK();
 
   const submitTx = useCallback(async () => {
     try {
@@ -16,18 +16,18 @@ const SafeApp = (): React.ReactElement => {
         txs: [
           {
             to: safe.safeAddress,
-            value: "0",
-            data: "0x",
+            value: '0',
+            data: '0x',
           },
         ],
-      })
-      console.log({ safeTxHash })
-      const safeTx = await sdk.txs.getBySafeTxHash(safeTxHash)
-      console.log({ safeTx })
+      });
+      console.log({ safeTxHash });
+      const safeTx = await sdk.txs.getBySafeTxHash(safeTxHash);
+      console.log({ safeTx });
     } catch (e) {
-      console.error(e)
+      console.error(e);
     }
-  }, [safe, sdk])
+  }, [safe, sdk]);
 
   return (
     <Container>
@@ -37,7 +37,7 @@ const SafeApp = (): React.ReactElement => {
         Click to send a test transaction
       </Button>
     </Container>
-  )
-}
+  );
+};
 
-export default SafeApp
+export default SafeApp;
