+++ packages/cra-template-safe-app/template/src/setupProxy.js
@@ -2,14 +2,13 @@
 // https://create-react-app.dev/docs/proxying-api-requests-in-development/#configuring-the-proxy-manually
 
 module.exports = function (app) {
-  app.use("/manifest.json", function (req, res, next) {
+  app.use('/manifest.json', function (req, res, next) {
     res.set({
-      "Access-Control-Allow-Origin": "*",
-      "Access-Control-Allow-Methods": "GET",
-      "Access-Control-Allow-Headers":
-        "X-Requested-With, content-type, Authorization",
-    })
+      'Access-Control-Allow-Origin': '*',
+      'Access-Control-Allow-Methods': 'GET',
+      'Access-Control-Allow-Headers': 'X-Requested-With, content-type, Authorization',
+    });
 
-    next()
-  })
-}
+    next();
+  });
+};
